new minor version: 1.3.0 - improve config.toml error printing
`toml` errors are more specific and they explicitely tell you the line,
the value and the possible values for the parameter in which you make a
typo. So instead of manually check and `.validate()`, let toml do it.

Also use types rather than "strings", they are safer given that the
first one generates a compiler error and the later one not when misstyped.

Thanks to Charles for the suggestions c: